+++ src/dispatch/mod.rs
@@ -3,7 +3,10 @@ pub use self::async_dispatcher::AsyncDispatcher;
 #[cfg(feature = "parallel")]
 pub use self::par_seq::{Par, ParSeq, RunWithPool, Seq};
 pub use self::{
-    batch::{BatchAccessor, BatchController, BatchUncheckedWorld, DefaultBatchControllerSystem},
+    batch::{
+        BatchAccessor, BatchController, BatchUncheckedWorld, MultiDispatchController,
+        MultiDispatcher,
+    },
     builder::DispatcherBuilder,
     dispatcher::Dispatcher,
 };