+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+* Batch dispatching ergonomics (getting rid of unsafe on the user side)
+  ([#197]).
+
 ## 0.10.2 (2020-02-13)
 
 ### Changed